ops/cdk/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.15.0
aws-cdk.aws-cloudwatch==1.15.0
aws-cdk.aws-ec2==1.15.0
aws-cdk.aws-events==1.15.0
aws-cdk.aws-iam==1.15.0
aws-cdk.aws-kms==1.15.0
aws-cdk.aws-lambda==1.15.0
aws-cdk.aws-logs==1.15.0
aws-cdk.aws-rds==1.15.0
aws-cdk.aws-s3==1.15.0
aws-cdk.aws-s3-assets==1.15.0
aws-cdk.aws-sam==1.15.0
aws-cdk.aws-secretsmanager==1.15.0
aws-cdk.aws-sqs==1.15.0
aws-cdk.aws-ssm==1.15.0
aws-cdk.core==1.15.0
aws-cdk.cx-api==1.15.0
aws-cdk.region-info==1.15.0
cattrs==0.9.0
importlib-resources==1.0.2
jsii==0.19.0
publication==0.0.3
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.4
